java(Array, Class)

+++ javabasic/src/task/prac.java
@@ -1,10 +1,7 @@
 package task;
 
-
 import java.util.Arrays;
 import java.util.Collections;
-
-
 
 public class prac {
     public static void main(String[] args) {
@@ -22,7 +19,7 @@ public class prac {
         // 이전 퍼즐을 다시 풀 때는 이전 퍼즐의 난이도에 상관없이 틀리지 않습니다
         // diff - level번 틀린 이후에 다시 퍼즐을 풀면 time_cur만큼의 시간을 사용하여 퍼즐을 해결합니다.
         // totalTime =  (diff - level) * (time_cur + time_prev) + time_cur
-        Long[] arrayDiffs = Arrays.stream(diffs).boxed().toArray(Long[]::new);
+        Integer[] arrayDiffs = Arrays.stream(diffs).boxed().toArray(Integer[]::new);
         Arrays.sort(arrayDiffs, Collections.reverseOrder());
         maxLevel = arrayDiffs[0];
         for(int posLevel = 1; posLevel < maxLevel; posLevel++) {
